EnlilGrids

EnlilGrids.gmsh_do_elements

EnlilGrids.gmsh_do_entities

EnlilGrids.gmsh_do_nodes

EnlilGrids.gmsh_do_physicalnames

EnlilGrids.import_gmsh

EnlilGrids.load_gmsh

EnlilGrids.load_gmsh_file

EnlilGrids.gmsh_do_elements — Method

gmsh_do_elements(raw_Elements) -> NamedTuple

Split raw_Elements created by load_gmsh_file.

Arguments

raw_Elements: data about elements in gmsh,

Keywords

Returns

NamedTuple: one Dict{Int,String} for each dimension, entityTag => physicalTag:

entityTag1D :: Array{Int64,1},

entityTag2D :: Array{Int64,1},

entityTag3D :: Array{Int64,1},

elementType1D :: Array{Int64,1},

elementType2D :: Array{Int64,1},

elementType3D :: Array{Int64,1},

elementTag1D :: Array{Int64,1},

elementTag2D :: Array{Int64,1},

elementTag3D :: Array{Int64,1},

nodeTags1D :: Array{Array{Int64,1},1},

nodeTags2D :: Array{Array{Int64,1},1},

nodeTags3D :: Array{Array{Int64,1},1}.

Throws

EnlilGrids.gmsh_do_entities — Method

gmsh_do_entities(raw_Entities) -> NamedTuple

Split raw_Entities created by load_gmsh_file in to one Dict for each type of entity. Operates with only one physicalTag per entity.

Arguments

raw_Entities: data about entities in gmsh,

Keywords

Returns

NamedTuple: one Dict{Int,String} for each dimension, entityTag => physicalTag:

pointTagtoPhysicalTag,

curveTagtoPhysicalTag,

surfaceTagtoPhysicalTag,

volumeTagtoPhysicalTag.

Throws

EnlilGrids.gmsh_do_nodes — Method

gmsh_do_nodes(raw_Nodes) -> NamedTuple

Split raw_Nodes created by load_gmsh_file. Operates with only non parametric curves.

Arguments

raw_Nodes: data about nodes in gmsh,

Keywords

Returns

NamedTuple: one Dict{Int,String} for each dimension, entityTag => physicalTag:

nodeTag::Array{Int64,1},

vx::Array{Float64,1},

vy::Array{Float64,1},

vz::Array{Float64,1}.

Throws

EnlilGrids.gmsh_do_physicalnames — Method

gmsh_do_physicalnames(raw_PhysicalNames) -> NamedTuple

Split raw_PhysicalNames created by load_gmsh_file in to one Dict for each dimension.

Arguments

raw_PhysicalNames: data about physical names in gmsh,

Keywords

Returns

NamedTuple: one Dict{Int,String} for each dimension, physicalTag => Name:

physicalTag1DtoName - maps physicalTag to Name for 1D objects,

physicalTag2DtoName - maps physicalTag to Name for 2D objects,

physicalTag3DtoName - maps physicalTag to Name for 3D objects,

Throws

EnlilGrids.import_gmsh — Method

import MeshCore ird0.

EnlilGrids.load_gmsh — Method

gmsh v 4.1

EnlilGrids.load_gmsh_file — Method

load_gmsh_file(filename::AbstractString) -> NamedTuple

Loads the filename into separete arrays according to the gmsh file format style in ASCII mode.

Arguments

filename::AbstractString: name of the file to be loaded,

Keywords

Returns

NamedTuple: gmsh file is separated into the arrays of substring:

msh_version - constains infomartion about gmsh file version,

msh_PhysicalNames - constains data about physical objects,

msh_Entities - constains data about entities,

msh_Nodes - constains data about nodes including coordinates,

msh_Elements - constains data about elements,

Throws

Error: file with name filename does not exists,

Error: file with name filename is saved in binary mode.